==> analysis_easifish/bydatasets/lt172/00.viz_s4-Copy2.ipynb <==
# Inputs
---

import numpy as np, zarr, nrrd
# alignment functions
from bigstream.align import alignment_pipeline
from bigstream.transform import apply_transform
import tifffile

import seaborn as sns
import matplotlib.pyplot as plt

p      = '/u/home/f/f7xiesnm/project-zipursky/data/hold/lt172/'
outdir = '/u/home/f/f7xiesnm/project-zipursky/easifish/results/viz_lt172'
!mkdir -p $outdir

!ls $p

# raw


rounds = ['r1', 'r2', 'r3', 'r4', 'r5']
for r in rounds:
    path = p + f'/{r}_export.n5'
    zarr_data = zarr.open(store=zarr.N5Store(path), mode='r')
    for c in ['c0', 'c1', 'c2', 'c3', 'c4']:
        img = zarr_data[f'/{c}/s4'] 
        print(np.max(img), np.min(img))
        output = outdir+f'/{r}_{c}_s4.tiff'
        print(output)
        tifffile.imwrite(output, img)

# reg

rounds = ['r1', 'r2', 'r4', 'r5']
for r in rounds:
    path = f'/u/home/f/f7xiesnm/project-zipursky/easifish/lt172/outputs/{r}/registration/{r}-to-r3/warped'
    
    zarr_data = zarr.open(store=zarr.N5Store(path), mode='r')
    img = zarr_data[f'/c3_reg/s3'][::2,::2,::2] 
    output = outdir+f'/{r}_c3_reg_s3d2x.tiff'
    print(output)
    tifffile.imwrite(output, img)
